SPRINT 18. VIRTUALIZATION. DOCKER IMAGE/docke_examples/example_django_harder/om-54-team_02/requirements.txt
asgiref==3.5.2
Django==4.1
# psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
djangorestframework==3.14.0
django-crispy-forms
django-environ==0.4.5